Conflicts:
	src/com/Boat/Empresa/Empresa.java
	src/com/Boat/Persona/Empleat.java

+++ src/com/Boat/Persona/Comercial.java
@@ -15,10 +15,12 @@ public class Comercial extends Empleat {
 
     private int comissio;
 
-    public Comercial(int comissio, Date dataAlta, int nomina, String nom, String llinatge, Document document, String numDocument, String adreça, int telefon, String email) {
+    public Comercial(int comissio, Date dataAlta, double nomina, String nom, String llinatge, Document document, String numDocument, String adreça, int telefon, String email) {
         super(dataAlta, nomina, nom, llinatge, document, numDocument, adreça, telefon, email);
         this.comissio = comissio;
     }
+
+
 
     public int getComissio() {
         return comissio;
